+++ src/main/java/com/nc/med/repo/SalesOrderRepo.java
@@ -19,7 +19,7 @@ public interface SalesOrderRepo extends JpaRepository<SalesOrder, Long> {
                                                                     LocalDateTime endDate, OrderStatus orderStatus);
 
     @Query(value = "SELECT previous_balance FROM SALES_ORDER where customer_id = :#{#customer.id} order by  sales_orderid desc limit 1", nativeQuery = true)
-    int findDueAmount(@Param("customer") Customer customer);
+    Double findDueAmount(@Param("customer") Customer customer);
 
     @Query(value = "select CAST(so.bill_date AS DATE), p.product_name, sum(sod.qty_ordered), sum(sod.sales_price), sum(sod.profit), sum(so.total_price)\n " +
             "from sales_order so, sales_order_detail sod, product p\n" +
